geotiff tests: consolidate nodata write-side cluster (Sub-PR B)
Cluster 10 of long-tail epic #2424. Fold 5 write-side nodata files
into a new ``xrspatial/geotiff/tests/write/test_nodata.py``:

- ``test_nodata_validation_1973.py`` (non-numeric / bool rejection)
- ``test_nodata_bool_rejection_1911.py`` (bool rejection at writer
  + ``build_geo_tags``)
- ``test_nodata_int64_precision_1847.py`` (full-precision 64-bit
  integer sentinels via ``_parse_nodata_str``; eager / dask / VRT /
  GPU round-trip)
- ``test_nodata_out_of_range_1581.py`` (out-of-range int sentinel
  is a value-match no-op rather than OverflowError)
- ``test_mask_nodata_kwarg_2052.py`` (``mask_nodata=False`` keeps
  source integer dtype so the ``dtype=`` cast contract is reachable)

Tests-only. The shared ``requires_gpu`` marker from
``_helpers/markers.py`` replaces the per-file ``_gpu_available`` /
``_gpu_only`` helpers. 87 tests collected, matching the baseline sum.

Refs #2434.
