+++ docs/source/reference/release_gate_geotiff.rst
@@ -387,7 +387,8 @@ Nodata lifecycle
        sentinels are preserved bit-exact, float sentinels surface as NaN
        only when ``mask_nodata=True``.
      - ``xrspatial/geotiff/tests/test_nodata_lifecycle_attrs_2135.py``,
-       ``xrspatial/geotiff/tests/test_nodata_lifecycle_parity_2211.py``
+       ``xrspatial/geotiff/tests/test_nodata_lifecycle_parity_2211.py``,
+       ``xrspatial/geotiff/tests/write/test_nodata.py``
      - `#2341`_
    * - ``attrs['masked_nodata']`` lifecycle signal
      - stable
